requirements.txt
youtube-transcript-api
langchain
langchain-core
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-huggingface
openai
google-generativeai
transformers
huggingface-hub
faiss-cpu
tiktoken
python-dotenv
numpy
scikit-learn
streamlit
